Refactor HyperLogLogPlusPlus to separate algorithms and internal data representation
